Pagination Bug, Add Record Structure & Centering, and a screenshot of errors
Updated the code structure in addRecord.php to have indented html instead of echo-ed html.  Now its easier to read with the indentation and coloring.  Also updated it to center buttons/images via bootstraps class instead of the basic html center tag.

Inserted some indentation in the pagination section of index.php to make it more readable.  Also fixed a small bug in it to wear the next button would disable one page too early.

Added a file with some screenshots of errors that occur after a simple browser refresh.  I need to handle a refresh on this screen a lot better than this.