add/edit image upload filename fixed for new filenames.  Changed some actual filenames.
Updated the add record and edit record so that the image uploaded successfully display the new file name if a new filename had to be created.  In the past it only ever displayed the original filename.

Added a file I can test in.

Changed test to zphpinfo (makes more sense) and changed related to displayRelated (also makes more sense and I like that the file name matches the displayItem since that file calls it anyway).